Updated constants for better SFCC organization
- Moved sorting and defaultSort definitions to lib/sfcc/constants.ts for improved structure.
- Updated import paths across multiple components to reflect the new constants location.
- Adjusted currency code handling in the carousel component for accuracy.
