+++ app/search/[collection]/page.tsx
@@ -3,8 +3,8 @@ import { notFound } from "next/navigation";
 
 import Grid from "components/grid";
 import ProductGridItems from "components/layout/product-grid-items";
-import { defaultSort, sorting } from "lib/constants";
 import { getCollection, getCollectionProducts } from "lib/sfcc";
+import { defaultSort, sorting } from "lib/sfcc/constants";
 
 export async function generateMetadata(props: {
   params: Promise<{ collection: string }>;
