+++ app/search/layout.tsx
@@ -1,15 +1,11 @@
-import Footer from 'components/layout/footer';
-import Collections from 'components/layout/search/collections';
-import FilterList from 'components/layout/search/filter';
-import { sorting } from 'lib/constants';
-import ChildrenWrapper from './children-wrapper';
-import { Suspense } from 'react';
+import Footer from "components/layout/footer";
+import Collections from "components/layout/search/collections";
+import FilterList from "components/layout/search/filter";
+import { sorting } from "lib/sfcc/constants";
+import { Suspense } from "react";
+import ChildrenWrapper from "./children-wrapper";
 
-export default function SearchLayout({
-  children
-}: {
-  children: React.ReactNode;
-}) {
+export default function SearchLayout({ children }: { children: React.ReactNode }) {
   return (
     <>
       <div className="mx-auto flex max-w-(--breakpoint-2xl) flex-col gap-8 px-4 pb-4 text-black md:flex-row dark:text-white">
