+++ components/layout/search/filter/index.tsx
@@ -1,4 +1,4 @@
-import { SortFilterItem } from "lib/constants";
+import { SortFilterItem } from "lib/sfcc/constants";
 import { Suspense } from "react";
 import FilterItemDropdown from "./dropdown";
 import { FilterItem } from "./item";
